HW03-TX: fixed chart display for react app
