+++ react_excercise/got-app/src/ChartData.js
@@ -1,0 +1,44 @@
+import { getHouseMembers } from './Api';
+
+const backgroundColors = [
+    'rgba(54, 162, 235, 0.8)',
+    'rgba(255, 206, 86, 0.8)',
+    'rgba(255, 99, 132, 0.8)',
+    'rgba(75, 192, 192, 0.8)',
+    'rgba(153, 102, 255, 0.8)',
+    'rgba(255, 159, 64, 0.8)',
+    'rgba(199, 199, 199, 0.8)',
+    'rgba(83, 102, 255, 0.8)',
+    'rgba(40, 159, 64, 0.8)',
+    'rgba(210, 199, 199, 0.8)',
+    'rgba(78, 52, 199, 0.8)',
+];
+
+const borderColors = [
+    'rgba(54, 162, 235, 1)',
+    'rgba(255, 206, 86, 1)',
+    'rgba(255, 99, 132, 1)',
+    'rgba(75, 192, 192, 1)',
+    'rgba(153, 102, 255, 1)',
+    'rgba(255, 159, 64, 1)',
+    'rgba(159, 159, 159, 1)',
+    'rgba(83, 102, 255, 1)',
+    'rgba(40, 159, 64, 1)',
+    'rgba(210, 199, 199, 1)',
+    'rgba(78, 52, 199, 1)',
+];
+const houseNames = await getHouseMembers();
+
+const chartData = {
+    labels: Object.keys(houseNames),
+    datasets: [
+        {
+            data: Object.values(houseNames),
+            backgroundColor: backgroundColors,
+            borderColor: borderColors,
+            hoverBackgroundColor: ['white'],
+        },
+    ],
+};
+
+export default chartData;
